ref #58 code adjusted as requested
